More code refactors

+++ subsys/greybus/platform/transport-dummy.c
@@ -4,13 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "transport.h"
+#include "../greybus_transport.h"
 
 static void init()
 {
 }
 
-static int listen(unsigned int cport)
+static int listen(uint16_t cport)
 {
 	return 0;
 }
@@ -20,7 +20,7 @@ static struct gb_transport_backend gb_trans_backend = {
 	.listen = listen,
 };
 
-struct gb_transport_backend *gb_transport_backend_init(size_t num_cports)
+struct gb_transport_backend *gb_transport_backend_init(uint16_t num_cports)
 {
 	return &gb_trans_backend;
 }
